Remove vec_to_tree implementation
We'll do it JS side instead. I tried to implement something like a tree,
and later found this SO answer:
https://stackoverflow.com/questions/33895090/serialize-json-in-a-recursive-struct
However right now I don't know how to implement it. In my case the
struct can contain other values than just a string, it can also contain
a child, so it's a real tree.

Actually, I just got an idea while writing this. What if, instead of a
struct, we use nested maps? So each HashMap can contain either an
HashMap or a value, in this case a Comment. It's worth looking into.